.cursor/skills/legacy-course-converter/scripts/requirements.txt
python-pptx==0.6.23
python-docx==1.1.2
openpyxl==3.1.5
pdfplumber==0.11.4
